* added shortcut function to create boxes  * rigid bodies shortcuts take an optional "motion_state" parameter to use custom classes instead of DefaultMotionState  * RigidBody.getMotionState() returns the python object directly rather than wrapping a new one from the underlying C++ object  * Transform.getOpenGLMatrix returns a float32 array

+++ ballistics/linearmath/transform.pyx
@@ -43,7 +43,7 @@ cdef public api class Transform[type BstxTransformType, object BstxTransform]:
     def getOpenGLMatrix(self):
         """Return the matrix as a 16 elements flat numpy array."""
         cdef btScalar m[16]
-        cdef np.ndarray ret = np.empty(16)
+        cdef np.ndarray ret = np.empty(16, dtype=np.float32)
         self.wrapped.getOpenGLMatrix(m)
         for i in range(16):
             ret[i] = m[i]
